refactor(tasks): move task API into feature
