+++ frontend/src/features/tasks/components/EditTaskModal.tsx
@@ -1,7 +1,7 @@
 import "../../../components/common.css"
 import "./EditTaskModal.css"
 import { useEffect, useState } from "react"
-import { updateTask } from "../../../lib/api"
+import { updateTask } from "../api/tasksApi"
 import type { Task } from "../types/task"
 import { useApiError } from "../../../hooks/useApiError";
 import { useNavigate } from "react-router-dom";
